requirements.txt
langchain
python-dotenv
ipykernel
tiktoken
langgraph
langchain_community
langchain-text-splitters
bs4
langchain-google-genai
google-generativeai
langchain-cohere
faiss-cpu
streamlit
langchain-groq
langchain-huggingface
sentence-transformers
langserve
wikipedia
langchainhub
duckduckgo-search
rich